requirements.txt
discord.py
python-dotenv
aiohttp
asyncpg
cryptography
PyYAML
fuzzywuzzy
python-Levenshtein
watchdog
timedelta
python-dotenv
requests
beautifulsoup4